close #8; moved preprocess_3h and preprocessing utils to pipeline.preprocess
