--- scripts/helpers/preprocess.py
@@ -1,20 +1,0 @@
-import datetime
-
-def categorize(x):
-    if x <= 30: return 0
-    elif x <= 50: return 1
-    else: return 2
-
-def preprocess_3h(df):
-
-    x = df.copy()
-
-    x['category'] = x['Kp*10'].apply(categorize)
-
-    x['dttm'] = x.apply(
-        lambda y: datetime.datetime(
-            int(y.year), int(y.month), 
-            int(y.day), int(y['hour from']), 0), axis=1)
-    
-    x_3h = x.sort_values(by='dttm').iloc[:: 3].bfill()
-    return x_3h
